added FileHandler for creating, highscore test in the appData directory. Also for reading and writing high scores

+++ SFML_Tetris/HighScores.h
@@ -2,9 +2,9 @@
 #include "SFML/Graphics.hpp"
 #include "DEFINITIONS.h"
 #include "IGuiElement.h"
+#include "FileHandler.h"
 #include <vector>
 #include <sstream>
-#include <fstream>
 
 class HighScores : public IGuiElement
 {
@@ -12,9 +12,9 @@ public:
 	HighScores() = default;
 
 	void Load();
-	void SetNew(const unsigned int newScore);
+	void SetNew(const size_t newScore);
 	void SetupTop10List(const sf::Font& font);
-	void AddCharToScoreHolder(const unsigned int keyCode);
+	void AddCharToScoreHolder(const size_t keyCode);
 	void RemoveCharFromScoreHolder();
 	void SaveScoreHolder();
 	bool IsNewScore() const;
@@ -27,11 +27,11 @@ public:
 	void SetGuiElementPositions() override;
 
 private:
-	std::vector<std::pair<std::string, unsigned int>> highScores;
+	std::vector<std::pair<std::string, size_t>> highScores;
 	std::vector<sf::Text> highScoreTexts;
 	sf::Text top10Text;
 
-	unsigned int newScoreIndex = 0;
+	size_t newScoreIndex = 0;
 	std::string scoreHolder = "";
 	bool scoreSet = false;
 	bool isNewScore = false;
@@ -41,7 +41,7 @@ private:
 
 	static constexpr float flashTime = 0.3f;
 	static constexpr char letters[27] = { "ABCDEFGHIJKLMNOPQRSTUVWXYZ" };
-	static constexpr unsigned int maxHighScoresSize = 10;
+	static constexpr size_t maxHighScoresSize = 10;
 
 };
